+++ lib/ex_aws/auth.ex
@@ -3,8 +3,12 @@ defmodule ExAws.Auth do
 
   alias ExAws.Auth.Credentials
   alias ExAws.Auth.Signatures
+  alias ExAws.Request.Url
 
   @moduledoc false
+
+  @unsignable_headers ["x-amzn-trace-id"]
+  @unsignable_headers_multi_case ["x-amzn-trace-id", "X-Amzn-Trace-Id"]
 
   def validate_config(config) do
     with :ok <- get_key(config, :secret_access_key),
@@ -71,34 +75,32 @@ defmodule ExAws.Auth do
       service = service_name(service)
       signed_headers = presigned_url_headers(url, headers)
 
-      org_query_params = query_params |> Enum.map(fn {k, v} -> {to_string(k), v} end)
+      uri = URI.parse(url)
+      uri_query = query_from_parsed_uri(uri)
+
+      org_query_params =
+        Enum.reduce(query_params, uri_query, fn {k, v}, acc -> [{to_string(k), v} | acc] end)
 
       amz_query_params =
         build_amz_query_params(service, datetime, config, expires, signed_headers)
 
-      [org_query, amz_query] =
-        [org_query_params, amz_query_params] |> Enum.map(&canonical_query_params/1)
+      query_to_sign = (org_query_params ++ amz_query_params) |> canonical_query_params()
 
-      query_to_sign = (org_query_params ++ amz_query_params) |> canonical_query_params
+      amz_query_string = canonical_query_params(amz_query_params)
 
       query_for_url =
-        if Enum.any?(org_query_params), do: org_query <> "&" <> amz_query, else: amz_query
-
-      uri = URI.parse(url)
-
-      path =
-        if uri.query do
-          uri.path <> "?" <> uri.query
+        if Enum.any?(org_query_params) do
+          canonical_query_params(org_query_params) <> "&" <> amz_query_string
         else
-          uri.path
+          amz_query_string
         end
 
-      path = uri_encode(path)
+      path = url |> Url.get_path(service) |> Url.uri_encode()
 
       signature =
         signature(
           http_method,
-          path,
+          url,
           query_to_sign,
           signed_headers,
           body,
@@ -119,15 +121,13 @@ defmodule ExAws.Auth do
   defp handle_temp_credentials(headers, _), do: headers
 
   defp auth_header(http_method, url, headers, body, service, datetime, config) do
-    uri = URI.parse(url)
-    path = uri_encode(uri.path)
-
     query =
-      if uri.query,
-        do: uri.query |> URI.decode_query() |> Enum.to_list() |> canonical_query_params,
-        else: ""
+      url
+      |> URI.parse()
+      |> query_from_parsed_uri()
+      |> canonical_query_params()
 
-    signature = signature(http_method, path, query, headers, body, service, datetime, config)
+    signature = signature(http_method, url, query, headers, body, service, datetime, config)
 
     [
       "AWS4-HMAC-SHA256 Credential=",
@@ -142,7 +142,16 @@ defmodule ExAws.Auth do
     |> IO.iodata_to_binary()
   end
 
-  defp signature(http_method, path, query, headers, body, service, datetime, config) do
+  defp query_from_parsed_uri(%{query: nil}), do: []
+
+  defp query_from_parsed_uri(%{query: query_string}) do
+    query_string
+    |> URI.decode_query()
+    |> Enum.to_list()
+  end
+
+  defp signature(http_method, url, query, headers, body, service, datetime, config) do
+    path = url |> Url.get_path(service) |> Url.uri_encode()
     request = build_canonical_request(http_method, path, query, headers, body)
     string_to_sign = string_to_sign(request, service, datetime, config)
     Signatures.generate_signature_v4(service, config, datetime, string_to_sign)
@@ -158,10 +167,7 @@ defmodule ExAws.Auth do
       |> Enum.map(fn {k, v} -> "#{k}:#{remove_dup_spaces(to_string(v))}" end)
       |> Enum.join("\n")
 
-    signed_headers_list =
-      headers
-      |> Keyword.keys()
-      |> Enum.join(";")
+    signed_headers_list = signed_headers_value(headers)
 
     payload =
       case body do
@@ -208,6 +214,7 @@ defmodule ExAws.Auth do
   defp signed_headers(headers) do
     headers
     |> Enum.map(fn {k, _} -> String.downcase(k) end)
+    |> Kernel.--(@unsignable_headers)
     |> Enum.sort(&(&1 < &2))
     |> Enum.join(";")
   end
@@ -216,9 +223,12 @@ defmodule ExAws.Auth do
 
   defp canonical_query_params(params) do
     params
-    |> Enum.sort(fn {k1, _}, {k2, _} -> k1 < k2 end)
+    |> Enum.sort(&compare_query_params/2)
     |> Enum.map_join("&", &pair/1)
   end
+
+  defp compare_query_params({key, value1}, {key, value2}), do: value1 < value2
+  defp compare_query_params({key_1, _}, {key_2, _}), do: key_1 < key_2
 
   defp pair({k, _}) when is_list(k) do
     raise ArgumentError, "encode_query/1 keys cannot be lists, got: #{inspect(k)}"
@@ -250,9 +260,10 @@ defmodule ExAws.Auth do
 
   defp canonical_headers(headers) do
     headers
-    |> Enum.map(fn
-      {k, v} when is_binary(v) -> {String.downcase(to_string(k)), String.trim(v)}
-      {k, v} -> {String.downcase(to_string(k)), v}
+    |> Enum.reduce([], fn
+      {k, _v}, acc when k in @unsignable_headers_multi_case -> acc
+      {k, v}, acc when is_binary(v) -> [{String.downcase(to_string(k)), String.trim(v)} | acc]
+      {k, v}, acc -> [{String.downcase(to_string(k)), v} | acc]
     end)
     |> Enum.sort(fn {k1, _}, {k2, _} -> k1 < k2 end)
   end
@@ -268,13 +279,19 @@ defmodule ExAws.Auth do
       {"X-Amz-Credential", Credentials.generate_credential_v4(service, config, datetime)},
       {"X-Amz-Date", amz_date(datetime)},
       {"X-Amz-Expires", expires},
-      {"X-Amz-SignedHeaders", Keyword.keys(signed_headers) |> Enum.join(";")}
+      {"X-Amz-SignedHeaders", signed_headers_value(signed_headers)}
     ] ++
       if config[:security_token] do
         [{"X-Amz-Security-Token", config[:security_token]}]
       else
         []
       end
+  end
+
+  defp signed_headers_value(headers) do
+    headers
+    |> Enum.map(&elem(&1, 0))
+    |> Enum.join(";")
   end
 
   defp service_override(service, config) do