+++ lib/ex_aws/config/defaults.ex
@@ -7,12 +7,13 @@ defmodule ExAws.Config.Defaults do
     access_key_id: [{:system, "AWS_ACCESS_KEY_ID"}, :instance_role],
     secret_access_key: [{:system, "AWS_SECRET_ACCESS_KEY"}, :instance_role],
     http_client: ExAws.Request.Hackney,
-    json_codec: Poison,
+    json_codec: Jason,
     retries: [
       max_attempts: 10,
       base_backoff_in_ms: 10,
       max_backoff_in_ms: 10_000
-    ]
+    ],
+    normalize_path: true
   }
 
   @doc """
@@ -41,6 +42,21 @@ defmodule ExAws.Config.Defaults do
   def defaults(:iot_data) do
     %{service_override: :iotdata}
     |> Map.merge(defaults(:iot))
+  end
+
+  def defaults(:"session.qldb") do
+    %{service_override: :qldb}
+    |> Map.merge(defaults(:qldb))
+  end
+
+  def defaults(:ingest_timestream) do
+    %{service_override: :timestream}
+    |> Map.merge(defaults(:timestream))
+  end
+
+  def defaults(:query_timestream) do
+    %{service_override: :timestream}
+    |> Map.merge(defaults(:timestream))
   end
 
   def defaults(_) do
@@ -83,6 +99,8 @@ defmodule ExAws.Config.Defaults do
   defp service_map(:lex_models), do: "models.lex"
   defp service_map(:dynamodb_streams), do: "streams.dynamodb"
   defp service_map(:iot_data), do: "data.iot"
+  defp service_map(:ingest_timestream), do: "ingest.timestream"
+  defp service_map(:query_timestream), do: "query.timestream"
 
   defp service_map(service) do
     service
@@ -90,11 +108,9 @@ defmodule ExAws.Config.Defaults do
     |> String.replace("_", "-")
   end
 
-  @partition_data :ex_aws
-                  |> :code.priv_dir()
-                  |> Path.join("endpoints.exs")
-                  |> File.read!()
-                  |> Code.eval_string()
+  @external_resource "priv/endpoints.exs"
+
+  @partition_data Code.eval_file("priv/endpoints.exs", File.cwd!())
                   |> elem(0)
                   |> Map.get("partitions")
                   |> Map.new(fn partition ->