+++ lib/ex_aws/credentials_ini.ex
@@ -3,13 +3,17 @@ if Code.ensure_loaded?(ConfigParser) do
     # as per https://docs.aws.amazon.com/cli/latest/topic/config-vars.html
     @valid_config_keys ~w(
       aws_access_key_id aws_secret_access_key aws_session_token region
-      role_arn source_profile credential_source external_id mfa_serial role_session_name
+      role_arn source_profile credential_source external_id mfa_serial role_session_name credential_process
     )
 
     def security_credentials(profile_name) do
       shared_credentials = profile_from_shared_credentials(profile_name)
       config_credentials = profile_from_config(profile_name)
       Map.merge(config_credentials, shared_credentials)
+    end
+
+    def parse_ini_file({:ok, contents}, :system) do
+      parse_ini_file({:ok, contents}, profile_name_from_env())
     end
 
     def parse_ini_file({:ok, contents}, profile_name) do
@@ -63,6 +67,7 @@ if Code.ensure_loaded?(ConfigParser) do
     defp profile_from_config(profile_name) do
       section =
         case profile_name do
+          :system -> "profile #{profile_name_from_env()}"
           "default" -> "default"
           other -> "profile #{other}"
         end
@@ -71,6 +76,10 @@ if Code.ensure_loaded?(ConfigParser) do
       |> Path.join(".aws/config")
       |> File.read()
       |> parse_ini_file(section)
+    end
+
+    defp profile_name_from_env() do
+      System.get_env("AWS_PROFILE") || "default"
     end
   end
 else