+++ priv/endpoints.exs
@@ -13,6 +13,7 @@
       "regions" => %{
         "ap-northeast-1" => %{"description" => "Asia Pacific (Tokyo)"},
         "ap-northeast-2" => %{"description" => "Asia Pacific (Seoul)"},
+        "ap-east-1" => %{"description" => "Asia Pacific (Hong Kong)"},
         "ap-south-1" => %{"description" => "Asia Pacific (Mumbai)"},
         "ap-southeast-1" => %{"description" => "Asia Pacific (Singapore)"},
         "ap-southeast-2" => %{"description" => "Asia Pacific (Sydney)"},
@@ -22,6 +23,7 @@
         "eu-west-2" => %{"description" => "EU (London)"},
         "eu-west-3" => %{"description" => "EU (Paris)"},
         "eu-north-1" => %{"description" => "EU (Stockholm)"},
+        "eu-south-1" => %{"description" => "EU (Milan)"},
         "sa-east-1" => %{"description" => "South America (Sao Paulo)"},
         "us-east-1" => %{"description" => "US East (N. Virginia)"},
         "us-east-2" => %{"description" => "US East (Ohio)"},
@@ -29,13 +31,27 @@
         "us-west-2" => %{"description" => "US West (Oregon)"}
       },
       "services" => %{
+        "connect" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-central-1" => %{},
+            "eu-west-2" => %{},
+            "us-east-1" => %{},
+            "us-west-2" => %{}
+          }
+        },
         "firehose" => %{
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
             "eu-central-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-3" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
             "us-west-1" => %{},
@@ -77,6 +93,7 @@
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
             "ap-south-1" => %{},
+            "ap-east-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "ca-central-1" => %{},
@@ -96,6 +113,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -167,6 +185,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -182,19 +201,37 @@
             "us-west-2" => %{}
           }
         },
+        "ingest.timestream" => %{
+          "endpoints" => %{
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{},
+            "eu-west-1" => %{}
+          }
+        },
+        "query.timestream" => %{
+          "endpoints" => %{
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{},
+            "eu-west-1" => %{}
+          }
+        },
         "translate" => %{
           "defaults" => %{"protocols" => ["https"]},
           "endpoints" => %{
             "eu-west-1" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
-            "us-west-2" => %{}
+            "us-west-2" => %{},
+            "eu-central-1" => %{}
           }
         },
         "acm" => %{
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -219,6 +256,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -250,6 +288,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -269,6 +308,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -309,6 +349,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -373,6 +414,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -446,6 +488,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -467,6 +510,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -525,6 +569,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -541,6 +586,15 @@
             "us-west-2" => %{}
           }
         },
+        "session.qldb" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "eu-west-1" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{}
+          }
+        },
         "shield" => %{
           "defaults" => %{
             "protocols" => ["https"],
@@ -554,6 +608,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -602,6 +657,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -683,6 +739,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -703,6 +760,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -723,6 +781,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -732,6 +791,7 @@
             "eu-west-2" => %{},
             "eu-west-3" => %{},
             "eu-north-1" => %{},
+            "eu-south-1" => %{},
             "sa-east-1" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
@@ -811,6 +871,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -841,6 +902,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -896,6 +958,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -961,6 +1024,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1013,6 +1077,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1031,9 +1096,20 @@
         },
         "servicediscovery" => %{
           "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ca-central-1" => %{},
+            "eu-central-1" => %{},
             "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "eu-west-3" => %{},
+            "sa-east-1" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
+            "us-west-1" => %{},
             "us-west-2" => %{}
           }
         },
@@ -1055,6 +1131,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1141,12 +1218,23 @@
         "rekognition" => %{
           "endpoints" => %{
             "ap-northeast-1" => %{},
+            "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
             "eu-west-1" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
             "us-west-2" => %{}
           }
+        },
+        "chime" => %{
+          "endpoints" => %{
+            "aws-global" => %{
+              "hostname" => "service.chime.aws.amazon.com",
+              "signatureVersions" => ["v2", "v3", "v4"]
+            }
+          },
+          "isRegionalized" => false,
+          "partitionEndpoint" => "aws-global"
         },
         "s3" => %{
           "defaults" => %{"protocols" => ["http", "https"], "signatureVersions" => ["s3v4"]},
@@ -1156,6 +1244,7 @@
               "signatureVersions" => ["s3", "s3v4"]
             },
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{
               "hostname" => "s3.ap-southeast-1.amazonaws.com",
@@ -1204,6 +1293,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1251,6 +1341,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1278,6 +1369,7 @@
               "credentialScope" => %{"region" => "ap-northeast-2"},
               "hostname" => "sts.ap-northeast-2.amazonaws.com"
             },
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1320,6 +1412,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1343,10 +1436,31 @@
         "comprehend" => %{
           "defaults" => %{"protocols" => ["https"]},
           "endpoints" => %{
+            "eu-central-1" => %{},
             "eu-west-1" => %{},
             "us-east-1" => %{},
             "us-east-2" => %{},
-            "us-west-2" => %{}
+            "us-west-2" => %{},
+            "ap-south-1" => %{},
+            "ap-northeast-2" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "ap-northeast-1" => %{},
+            "ca-central-1" => %{},
+            "us-gov-west-1" => %{}
+          }
+        },
+        "textract" => %{
+          "defaults" => %{"protocols" => ["https"]},
+          "endpoints" => %{
+            "us-east-2" => %{},
+            "us-east-1" => %{},
+            "us-west-1" => %{},
+            "us-west-2" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{}
           }
         },
         "waf" => %{
@@ -1387,6 +1501,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1407,6 +1522,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1423,7 +1539,22 @@
             "us-west-2" => %{}
           }
         },
-        "email" => %{"endpoints" => %{"eu-west-1" => %{}, "us-east-1" => %{}, "us-west-2" => %{}}},
+        "email" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "ap-south-1" => %{},
+            "ap-southeast-1" => %{},
+            "ap-southeast-2" => %{},
+            "eu-central-1" => %{},
+            "eu-west-1" => %{},
+            "eu-west-2" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{},
+            "ca-central-1" => %{},
+            "sa-east-1" => %{}
+          }
+        },
         "storagegateway" => %{
           "endpoints" => %{
             "ap-northeast-1" => %{},
@@ -1452,6 +1583,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1472,6 +1604,7 @@
           "defaults" => %{"credentialScope" => %{"service" => "cloudhsm"}},
           "endpoints" => %{
             "ap-northeast-1" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1624,6 +1757,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1644,6 +1778,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1679,6 +1814,7 @@
         "route53domains" => %{"endpoints" => %{"us-east-1" => %{}}},
         "states" => %{
           "endpoints" => %{
+            "ap-east-1" => %{},
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
             "ap-southeast-1" => %{},
@@ -1697,6 +1833,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1818,6 +1955,7 @@
           "endpoints" => %{
             "ap-northeast-1" => %{},
             "ap-northeast-2" => %{},
+            "ap-east-1" => %{},
             "ap-south-1" => %{},
             "ap-southeast-1" => %{},
             "ap-southeast-2" => %{},
@@ -1907,6 +2045,15 @@
           },
           "isRegionalized" => false,
           "partitionEndpoint" => "aws-global"
+        },
+        "qldb" => %{
+          "endpoints" => %{
+            "ap-northeast-1" => %{},
+            "eu-west-1" => %{},
+            "us-east-1" => %{},
+            "us-east-2" => %{},
+            "us-west-2" => %{}
+          }
         }
       }
     },
@@ -1988,6 +2135,7 @@
           "defaults" => %{"credentialScope" => %{"service" => "execute-api"}},
           "endpoints" => %{"cn-north-1" => %{}}
         },
+        "chime" => %{"endpoints" => %{"defaults" => %{}}},
         "s3" => %{
           "defaults" => %{"protocols" => ["http", "https"], "signatureVersions" => ["s3v4"]},
           "endpoints" => %{"cn-north-1" => %{}, "cn-northwest-1" => %{}}
@@ -2041,10 +2189,13 @@
       "partition" => "aws-us-gov",
       "partitionName" => "AWS GovCloud (US)",
       "regionRegex" => "^us\\-gov\\-\\w+\\-\\d+$",
-      "regions" => %{"us-gov-west-1" => %{"description" => "AWS GovCloud (US)"}},
+      "regions" => %{
+        "us-gov-east-1" => %{"description" => "AWS GovCloud (US-East)"},
+        "us-gov-west-1" => %{"description" => "AWS GovCloud (US-West)"}
+      },
       "services" => %{
-        "elasticache" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "config" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
+        "elasticache" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "config" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "iam" => %{
           "endpoints" => %{
             "aws-us-gov-global" => %{
@@ -2055,62 +2206,110 @@
           "isRegionalized" => false,
           "partitionEndpoint" => "aws-us-gov-global"
         },
-        "snowball" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "glacier" => %{"endpoints" => %{"us-gov-west-1" => %{"protocols" => ["http", "https"]}}},
-        "acm" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "elasticmapreduce" => %{
-          "endpoints" => %{"us-gov-west-1" => %{"protocols" => ["http", "https"]}}
+        "snowball" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "glacier" => %{
+          "endpoints" => %{
+            "us-gov-east-1" => %{"protocols" => ["http", "https"]},
+            "us-gov-west-1" => %{"protocols" => ["http", "https"]}
+          }
         },
-        "logs" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "kinesis" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "events" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "sns" => %{"endpoints" => %{"us-gov-west-1" => %{"protocols" => ["http", "https"]}}},
+        "acm" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "elasticmapreduce" => %{
+          "endpoints" => %{
+            "us-gov-east-1" => %{"protocols" => ["http", "https"]},
+            "us-gov-west-1" => %{"protocols" => ["http", "https"]}
+          }
+        },
+        "logs" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "kinesis" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "events" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "sns" => %{
+          "endpoints" => %{
+            "us-gov-east-1" => %{"protocols" => ["http", "https"]},
+            "us-gov-west-1" => %{"protocols" => ["http", "https"]}
+          }
+        },
         "polly" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "lambda" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
+        "lambda" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "dynamodb" => %{
           "endpoints" => %{
-            "us-gov-west-1" => %{},
+            "us-gov-east-1-fips" => %{
+              "credentialScope" => %{"region" => "us-gov-east-1"},
+              "hostname" => "dynamodb.us-gov-east-1.amazonaws.com"
+            },
             "us-gov-west-1-fips" => %{
               "credentialScope" => %{"region" => "us-gov-west-1"},
               "hostname" => "dynamodb.us-gov-west-1.amazonaws.com"
-            }
+            },
+            "us-gov-east-1" => %{},
+            "us-gov-west-1" => %{}
           }
         },
         "streams.dynamodb" => %{
           "defaults" => %{"credentialScope" => %{"service" => "dynamodb"}},
           "endpoints" => %{
-            "us-gov-west-1" => %{},
+            "us-gov-east-1-fips" => %{
+              "credentialScope" => %{"region" => "us-gov-east-1"},
+              "hostname" => "dynamodb.us-gov-east-1.amazonaws.com"
+            },
             "us-gov-west-1-fips" => %{
               "credentialScope" => %{"region" => "us-gov-west-1"},
               "hostname" => "dynamodb.us-gov-west-1.amazonaws.com"
-            }
+            },
+            "us-gov-east-1" => %{},
+            "us-gov-west-1" => %{}
           }
         },
-        "es" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "monitoring" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
+        "es" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "monitoring" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "elasticloadbalancing" => %{
-          "endpoints" => %{"us-gov-west-1" => %{"protocols" => ["http", "https"]}}
+          "endpoints" => %{
+            "us-gov-east-1" => %{"protocols" => ["http", "https"]},
+            "us-gov-west-1" => %{"protocols" => ["http", "https"]}
+          }
         },
         "metering.marketplace" => %{
           "defaults" => %{"credentialScope" => %{"service" => "aws-marketplace"}},
-          "endpoints" => %{"us-gov-west-1" => %{}}
+          "endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}
         },
-        "ec2" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "apigateway" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "kms" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "sms" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "rds" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "cloudtrail" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "ssm" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "ecr" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
+        "ec2" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "apigateway" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "kms" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "sms" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "rds" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "cloudtrail" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "ssm" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "ecr" => %{
+          "endpoints" => %{
+            "fips-us-gov-east-1" => %{
+              "credentialScope" => %{"region" => "us-gov-east-1"},
+              "hostname" => "ecr-fips.us-gov-east-1.amazonaws.com"
+            },
+            "fips-us-gov-west_fips-1" => %{
+              "credentialScope" => %{"region" => "us-gov-west-1"},
+              "hostname" => "ecr-fips.us-gov-west-1.amazonaws.com"
+            },
+            "us-gov-east-1" => %{"protocols" => ["http", "https"]},
+            "us-gov-west-1" => %{"protocols" => ["http", "https"]}
+          }
+        },
         "cloudhsm" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
         "rekognition" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
+        "chime" => %{"endpoints" => %{"defaults" => %{}}},
         "s3" => %{
           "defaults" => %{"signatureVersions" => ["s3", "s3v4"]},
           "endpoints" => %{
+            "fips-us-gov-east-1" => %{
+              "credentialScope" => %{"region" => "us-gov-east-1"},
+              "hostname" => "s3-fips.us-gov-east-1.amazonaws.com"
+            },
             "fips-us-gov-west-1" => %{
               "credentialScope" => %{"region" => "us-gov-west-1"},
-              "hostname" => "s3-fips-us-gov-west-1.amazonaws.com"
+              "hostname" => "s3-fips.us-gov-west-1.amazonaws.com"
+            },
+            "us-gov-east-1" => %{
+              "hostname" => "s3.us-gov-east-1.amazonaws.com",
+              "protocols" => ["http", "https"]
             },
             "us-gov-west-1" => %{
               "hostname" => "s3.us-gov-west-1.amazonaws.com",
@@ -2118,32 +2317,53 @@
             }
           }
         },
-        "elasticbeanstalk" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "sts" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
+        "elasticbeanstalk" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "sts" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "sqs" => %{
           "endpoints" => %{
+            "us-gov-east-1" => %{
+              "protocols" => ["http", "https"],
+              "sslCommonName" => "{region}.queue.{dnsSuffix}"
+            },
             "us-gov-west-1" => %{
               "protocols" => ["http", "https"],
               "sslCommonName" => "{region}.queue.{dnsSuffix}"
             }
           }
         },
-        "tagging" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "redshift" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "storagegateway" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
+        "tagging" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "redshift" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "storagegateway" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
         "autoscaling" => %{
-          "endpoints" => %{"us-gov-west-1" => %{"protocols" => ["http", "https"]}}
+          "endpoints" => %{
+            "us-gov-east-1" => %{"protocols" => ["http", "https"]},
+            "us-gov-west-1" => %{"protocols" => ["http", "https"]}
+          }
         },
         "cloudhsmv2" => %{
           "defaults" => %{"credentialScope" => %{"service" => "cloudhsm"}},
-          "endpoints" => %{"us-gov-west-1" => %{}}
+          "endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}
         },
-        "dms" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "codedeploy" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "ecs" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "directconnect" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "cloudformation" => %{"endpoints" => %{"us-gov-west-1" => %{}}},
-        "swf" => %{"endpoints" => %{"us-gov-west-1" => %{}}}
+        "dms" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "codedeploy" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "ecs" => %{
+          "endpoints" => %{
+            "fips-us-gov-east-1" => %{
+              "credentialScope" => %{"region" => "us-gov-east-1"},
+              "hostname" => "ecs-fips-us-gov-east-1.amazonaws.com"
+            },
+            "fips-us-gov-west-1" => %{
+              "credentialScope" => %{"region" => "us-gov-west-1"},
+              "hostname" => "ecs-fips-us-gov-west-1.amazonaws.com"
+            },
+            "us-gov-east-1" => %{"protocols" => ["http", "https"]},
+            "us-gov-west-1" => %{"protocols" => ["http", "https"]}
+          }
+        },
+        "directconnect" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "cloudformation" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "swf" => %{"endpoints" => %{"us-gov-east-1" => %{}, "us-gov-west-1" => %{}}},
+        "sagemaker" => %{"endpoints" => %{"us-gov-west-1" => %{}}}
       }
     }
   ],